src/sports_prediction/data_collection/web_scraper.py
"""Web scraper for additional sports data sources."""

import asyncio
from typing import Dict, List, Any, Optional
from datetime import datetime
import aiohttp
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from ..utils.cache import CacheManager
from ..utils.rate_limiter import RateLimiter
from ..utils.logger import get_logger
from ..config.settings import settings

logger = get_logger(__name__)


class WebScraper:
    """Web scraper for sports data from various websites."""
    
    def __init__(self):
        self.cache = CacheManager()
        self.rate_limiter = RateLimiter()
        self.session: Optional[aiohttp.ClientSession] = None
        
        # Chrome options for Selenium
        self.chrome_options = Options()
        self.chrome_options.add_argument("--headless")
        self.chrome_options.add_argument("--no-sandbox")
        self.chrome_options.add_argument("--disable-dev-shm-usage")
        self.chrome_options.add_argument("--disable-gpu")
        self.chrome_options.add_argument("--window-size=1920,1080")
        
        # Target websites
        self.scrapers = {
            'transfermarkt': self._scrape_transfermarkt,
            'fbref': self._scrape_fbref,
            'basketball_reference': self._scrape_basketball_reference,
            'pro_football_reference': self._scrape_pro_football_reference,
            'hockey_reference': self._scrape_hockey_reference,
        }
    
    async def __aenter__(self):
        """Async context manager entry."""
        self.session = aiohttp.ClientSession(
            timeout=aiohttp.ClientTimeout(total=30),
            headers={
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
            }
        )
        return self
    
    async def __aexit__(self, exc_type, exc_val, exc_tb):
        """Async context manager exit."""
        if self.session:
            await self.session.close()
    
    async def scrape_team_data(self, source: str, sport: str, team_name: str) -> Dict[str, Any]:
        """Scrape team data from specified source."""
        if source not in self.scrapers:
            logger.warning(f"Scraper for {source} not implemented")
            return {}
        
        cache_key = f"scrape:{source}:{sport}:{team_name}"
        cached_data = self.cache.get(cache_key)
        if cached_data:
            return cached_data
        
        try:
            data = await self.scrapers[source](sport, team_name, 'team')
            if data:
                self.cache.set(cache_key, data, ttl=3600)  # Cache for 1 hour
            return data
        except Exception as e:
            logger.error(f"Error scraping {source} for {team_name}: {e}")
            return {}
    
    async def scrape_player_data(self, source: str, sport: str, player_name: str) -> Dict[str, Any]:
        """Scrape player data from specified source."""
        if source not in self.scrapers:
            logger.warning(f"Scraper for {source} not implemented")
            return {}
        
        cache_key = f"scrape:{source}:{sport}:{player_name}"
        cached_data = self.cache.get(cache_key)
        if cached_data:
            return cached_data
        
        try:
            data = await self.scrapers[source](sport, player_name, 'player')
            if data:
                self.cache.set(cache_key, data, ttl=3600)
            return data
        except Exception as e:
            logger.error(f"Error scraping {source} for {player_name}: {e}")
            return {}
    
    async def _scrape_transfermarkt(self, sport: str, name: str, data_type: str) -> Dict[str, Any]:
        """Scrape data from Transfermarkt (soccer)."""
        if sport not in ['mls', 'premier_league', 'la_liga', 'bundesliga', 'serie_a', 'champions_league']:
            return {}
        
        self.rate_limiter.wait_if_needed('transfermarkt')
        
        # Search for the team/player
        search_url = f"https://www.transfermarkt.com/schnellsuche/ergebnis/schnellsuche"
        search_params = {'query': name}
        
        if not self.session:
            return {}
        
        try:
            async with self.session.get(search_url, params=search_params) as response:
                if response.status != 200:
                    return {}
                
                html = await response.text()
                soup = BeautifulSoup(html, 'html.parser')
                
                # Parse search results and get the first relevant result
                if data_type == 'team':
                    return self._parse_transfermarkt_team(soup, name)
                else:
                    return self._parse_transfermarkt_player(soup, name)
        
        except Exception as e:
            logger.error(f"Error scraping Transfermarkt: {e}")
            return {}
    
    async def _scrape_fbref(self, sport: str, name: str, data_type: str) -> Dict[str, Any]:
        """Scrape data from FBRef (soccer statistics)."""
        if sport not in ['mls', 'premier_league', 'la_liga', 'bundesliga', 'serie_a', 'champions_league']:
            return {}
        
        self.rate_limiter.wait_if_needed('fbref')
        
        # FBRef search
        search_url = f"https://fbref.com/en/search/search.fcgi"
        search_params = {'search': name}
        
        if not self.session:
            return {}
        
        try:
            async with self.session.get(search_url, params=search_params) as response:
                if response.status != 200:
                    return {}
                
                html = await response.text()
                soup = BeautifulSoup(html, 'html.parser')
                
                if data_type == 'team':
                    return self._parse_fbref_team(soup, name)
                else:
                    return self._parse_fbref_player(soup, name)
        
        except Exception as e:
            logger.error(f"Error scraping FBRef: {e}")
            return {}
    
    async def _scrape_basketball_reference(self, sport: str, name: str, data_type: str) -> Dict[str, Any]:
        """Scrape data from Basketball Reference (NBA)."""
        if sport != 'nba':
            return {}
        
        self.rate_limiter.wait_if_needed('basketball_reference')
        
        # Basketball Reference search
        search_url = f"https://www.basketball-reference.com/search/search.fcgi"
        search_params = {'search': name}
        
        if not self.session:
            return {}
        
        try:
            async with self.session.get(search_url, params=search_params) as response:
                if response.status != 200:
                    return {}
                
                html = await response.text()
                soup = BeautifulSoup(html, 'html.parser')
                
                if data_type == 'team':
                    return self._parse_basketball_ref_team(soup, name)
                else:
                    return self._parse_basketball_ref_player(soup, name)
        
        except Exception as e:
            logger.error(f"Error scraping Basketball Reference: {e}")
            return {}
    
    async def _scrape_pro_football_reference(self, sport: str, name: str, data_type: str) -> Dict[str, Any]:
        """Scrape data from Pro Football Reference (NFL)."""
        if sport != 'nfl':
            return {}
        
        # Similar implementation to basketball reference
        return {}
    
    async def _scrape_hockey_reference(self, sport: str, name: str, data_type: str) -> Dict[str, Any]:
        """Scrape data from Hockey Reference (NHL)."""
        if sport != 'nhl':
            return {}
        
        # Similar implementation to basketball reference
        return {}
    
    def _parse_transfermarkt_team(self, soup: BeautifulSoup, team_name: str) -> Dict[str, Any]:
        """Parse Transfermarkt team data."""
        # Implementation would parse team market value, squad info, etc.
        return {
            'source': 'transfermarkt',
            'team_name': team_name,
            'market_value': None,
            'squad_size': None,
            'average_age': None,
            'foreigners': None
        }
    
    def _parse_transfermarkt_player(self, soup: BeautifulSoup, player_name: str) -> Dict[str, Any]:
        """Parse Transfermarkt player data."""
        # Implementation would parse player market value, transfer history, etc.
        return {
            'source': 'transfermarkt',
            'player_name': player_name,
            'market_value': None,
            'position': None,
            'age': None,
            'contract_expires': None
        }
    
    def _parse_fbref_team(self, soup: BeautifulSoup, team_name: str) -> Dict[str, Any]:
        """Parse FBRef team statistics."""
        # Implementation would parse detailed team statistics
        return {
            'source': 'fbref',
            'team_name': team_name,
            'goals_for': None,
            'goals_against': None,
            'xg_for': None,
            'xg_against': None,
            'possession': None
        }
    
    def _parse_fbref_player(self, soup: BeautifulSoup, player_name: str) -> Dict[str, Any]:
        """Parse FBRef player statistics."""
        # Implementation would parse detailed player statistics
        return {
            'source': 'fbref',
            'player_name': player_name,
            'goals': None,
            'assists': None,
            'xg': None,
            'xa': None,
            'minutes': None
        }
    
    def _parse_basketball_ref_team(self, soup: BeautifulSoup, team_name: str) -> Dict[str, Any]:
        """Parse Basketball Reference team statistics."""
        return {
            'source': 'basketball_reference',
            'team_name': team_name,
            'points_per_game': None,
            'rebounds_per_game': None,
            'assists_per_game': None,
            'field_goal_percentage': None
        }
    
    def _parse_basketball_ref_player(self, soup: BeautifulSoup, player_name: str) -> Dict[str, Any]:
        """Parse Basketball Reference player statistics."""
        return {
            'source': 'basketball_reference',
            'player_name': player_name,
            'points_per_game': None,
            'rebounds_per_game': None,
            'assists_per_game': None,
            'field_goal_percentage': None
        }
