requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
lightgbm>=4.0.0
xgboost>=1.7.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
aiohttp>=3.8.0
scrapy>=2.9.0

# Telegram Bot
# Core dependencies
python-telegram-bot==20.7
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1

# Data processing
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# Machine Learning - PyTorch
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Machine Learning - TensorFlow
tensorflow==2.15.0
tensorflow-probability==0.22.1

# Machine Learning - Tree-based models
lightgbm==4.1.0
xgboost==2.0.3
catboost==1.2.2

# Machine Learning - Scikit-learn
scikit-learn==1.3.2
scikit-optimize==0.9.0

# Machine Learning - Additional
optuna==3.5.0
hyperopt==0.2.7
shap==0.44.0

# Data collection and APIs
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
lxml==4.9.4

# Database and caching
redis==5.0.1
pymongo==4.6.1
motor==3.3.2
sqlalchemy==2.0.25
alembic==1.13.1

# AWS services
boto3==1.34.34
botocore==1.34.34

# Configuration and settings
pydantic==2.5.3
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# Web framework (for webhooks and API)
fastapi==0.108.0
uvicorn==0.25.0
gunicorn==21.2.0

# Data validation and serialization
marshmallow==3.20.2
jsonschema==4.20.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
click==8.1.7
tqdm==4.66.1
joblib==1.3.2

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# HTTP client improvements
httpx==0.26.0
urllib3==2.1.0

# Data visualization (optional)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Time series analysis
statsmodels==0.14.1
pmdarima==2.0.4

# Natural language processing (for text analysis)
nltk==3.8.1
textblob==0.17.1

# Image processing (for logos, charts)
Pillow==10.1.0

# Parallel processing
multiprocessing-logging==0.3.4
concurrent-futures==3.1.1

# Memory optimization
psutil==5.9.7
memory-profiler==0.61.0

# Configuration management
hydra-core==1.3.2
omegaconf==2.3.0

# API rate limiting
ratelimit==2.2.1
backoff==2.2.1

# Data compression
lz4==4.3.3
zstandard==0.22.0

# Caching utilities
diskcache==5.6.3
cachetools==5.3.2

# Async utilities
asyncio-throttle==1.0.2
aioredis==2.0.1

# Financial calculations (for betting odds)
QuantLib==1.32

# Geographic data (for venue information)
geopy==2.4.1

# Email notifications (optional)
sendgrid==6.11.0

# Webhook validation
hmac==20101005>=20.0
asyncio-telegram>=0.1.0

# AWS Services
boto3>=1.28.0
botocore>=1.31.0

# Database
pymongo>=4.4.0
redis>=4.6.0
sqlalchemy>=2.0.0

# Data Processing
polars>=0.18.0
pyarrow>=12.0.0
dask>=2023.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# API and Web
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0

# Configuration and Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0
